+++ data-plane-aas/src/main/java/de/fraunhofer/iosb/dataplane/aas/pipeline/AasDataSource.java
@@ -13,14 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.app.dataplane.aas.pipeline;
+package de.fraunhofer.iosb.dataplane.aas.pipeline;
 
-import de.fraunhofer.iosb.app.util.HttpRestClient;
-import okhttp3.Headers;
+import de.fraunhofer.iosb.aas.AasReader;
+import de.fraunhofer.iosb.dataplane.aas.spi.AasDataAddress;
 import okhttp3.MediaType;
-import okhttp3.Request;
 import okhttp3.ResponseBody;
-import org.eclipse.edc.connector.dataplane.http.pipeline.HttpPart;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSource;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.StreamResult;
 import org.eclipse.edc.spi.EdcException;
@@ -28,31 +26,28 @@ import org.eclipse.edc.spi.monitor.Monitor;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 
 import static java.lang.String.format;
-import static org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress.OCTET_STREAM;
 
 /**
  * Data source for new FA³ST with possibly self-signed certificate.
- * Inspired by {@link org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress}
+ * Inspired by HttpDataAddress
  */
 public class AasDataSource implements DataSource {
 
     private static final int FORBIDDEN = 401;
     private static final int NOT_AUTHORIZED = 403;
     private static final int NOT_FOUND = 404;
+    private static final String APPLICATION_JSON = "application/json";
 
     private String requestId;
     private Monitor monitor;
-    private HttpRestClient httpClient;
-    private String baseUrl;
-    private String path;
-    private Map<String, String> headers;
+    private AasReader aasReader;
+    private AasDataAddress aasDataAddress;
     private final AtomicReference<ResponseBodyStream> responseBodyStream = new AtomicReference<>();
 
     private AasDataSource() {
@@ -60,15 +55,10 @@ public class AasDataSource implements DataSource {
 
     @Override
     public StreamResult<Stream<Part>> openPartStream() {
-        var request = new Request.Builder()
-                .url(baseUrl + path)
-                .headers(Headers.of(headers))
-                .get()
-                .build();
 
         try {
             // NB: Do not close the response as the body input stream needs to be read after this method returns. The response closes the body stream.
-            var response = httpClient.execute(request);
+            var response = aasReader.get(aasDataAddress);
 
             if (response.isSuccessful()) {
                 var body = response.body();
@@ -77,8 +67,8 @@ public class AasDataSource implements DataSource {
                 }
                 var bodyStream = body.byteStream();
                 responseBodyStream.set(new ResponseBodyStream(body, bodyStream));
-                var mediaType = Optional.ofNullable(body.contentType()).map(MediaType::toString).orElse(OCTET_STREAM);
-                return StreamResult.success(Stream.of(new HttpPart("AAS Part", bodyStream, mediaType)));
+                var mediaType = Optional.ofNullable(body.contentType()).map(MediaType::toString).orElse(APPLICATION_JSON);
+                return StreamResult.success(Stream.of(new AasPart("AAS Part", bodyStream, mediaType)));
 
             } else {
                 try {
@@ -135,8 +125,8 @@ public class AasDataSource implements DataSource {
             return this;
         }
 
-        public Builder httpClient(HttpRestClient httpClient) {
-            dataSource.httpClient = httpClient;
+        public Builder aasReader(AasReader httpClient) {
+            dataSource.aasReader = httpClient;
             return this;
         }
 
@@ -145,27 +135,18 @@ public class AasDataSource implements DataSource {
             return this;
         }
 
-        public Builder baseUrl(String baseUrl) {
-            dataSource.baseUrl = baseUrl;
+        public Builder aasDataAddress(AasDataAddress aasDataAddress) {
+            dataSource.aasDataAddress = aasDataAddress;
             return this;
         }
 
-        public Builder path(String urlPath) {
-            dataSource.path = urlPath;
-            return this;
-        }
-
-        public Builder headers(Map<String, String> headers) {
-            dataSource.headers = headers;
-            return this;
-        }
 
         public AasDataSource build() {
             Objects.requireNonNull(dataSource.requestId, "requestId");
-            Objects.requireNonNull(dataSource.httpClient, "httpClient");
+            Objects.requireNonNull(dataSource.aasReader, "httpClient");
             Objects.requireNonNull(dataSource.monitor, "monitor");
-            Objects.requireNonNull(dataSource.baseUrl, "baseUrl");
-            Objects.requireNonNull(dataSource.path, "path");
+            Objects.requireNonNull(dataSource.aasDataAddress, "aasDataAddress");
+
             return dataSource;
         }