+++ data-plane-aas/src/main/java/de/fraunhofer/iosb/dataplane/aas/spi/AasDataAddress.java
@@ -13,36 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.app.dataplane.aas.pipeline;
+package de.fraunhofer.iosb.dataplane.aas.spi;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import de.fraunhofer.iosb.app.util.Encoder;
+import de.fraunhofer.iosb.util.Encoder;
 import org.eclipse.digitaltwin.aas4j.v3.model.Reference;
 import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultReference;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.types.domain.DataAddress;
 
+import java.util.Collections;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 
-import static de.fraunhofer.iosb.app.dataplane.aas.pipeline.AasDataSourceFactory.AAS_DATA_TYPE;
+import static de.fraunhofer.iosb.dataplane.aas.pipeline.AasDataSourceFactory.AAS_DATA_TYPE;
 import static java.lang.String.format;
 import static java.util.stream.Collectors.toMap;
-import static org.eclipse.edc.dataaddress.httpdata.spi.HttpDataAddressSchema.BASE_URL;
+
 
 /**
- * Inspired by {@link org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress}
+ * Inspired by  org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress
  * Enables more specific communication with AAS services
  */
-@JsonTypeName()
+@JsonTypeName
 @JsonDeserialize(builder = DataAddress.Builder.class)
 public class AasDataAddress extends DataAddress {
 
     public static final String REFERENCE_CHAIN = "referenceChain";
+    public static final String BASE_URL = "https://w3id.org/edc/v0.0.1/ns/baseUrl";
 
     private static final String METHOD = "method";
     private static final String ADDITIONAL_HEADER = "header";
@@ -55,6 +58,10 @@ public class AasDataAddress extends DataAddress {
     @JsonIgnore
     public String getBaseUrl() {
         return getStringProperty(BASE_URL);
+    }
+
+    public String getMethod() {
+        return getStringProperty(METHOD);
     }
 
     public Map<String, String> getAdditionalHeaders() {
@@ -141,10 +148,16 @@ public class AasDataAddress extends DataAddress {
             return this;
         }
 
+        public Builder copyFrom(DataAddress other) {
+            (Optional.ofNullable(other).map(DataAddress::getProperties).orElse(Collections.emptyMap())).forEach(this::property);
+            return this;
+        }
+
         @Override
         public AasDataAddress build() {
             this.type(AAS_DATA_TYPE);
             return address;
         }
     }
+
 }